test: extract insertion-fill unit tests to unit/tracks/
